[Refactor] Move non fungible jsons [erc1155, erc271] fetchers to separate files #5030
